rename token.py -> token_class.py
