+++ allennlp/data/__init__.py
@@ -9,7 +9,6 @@ from allennlp.data.fields.text_field import TextFieldTensors
 from allennlp.data.instance import Instance
 from allennlp.data.samplers import BatchSampler, Sampler
 from allennlp.data.token_indexers.token_indexer import TokenIndexer, IndexedTokenList
-from allennlp.data.tokenizers.token import Token
-from allennlp.data.tokenizers.tokenizer import Tokenizer
+from allennlp.data.tokenizers import Token, Tokenizer
 from allennlp.data.vocabulary import Vocabulary
 from allennlp.data.batch import Batch
