+++ allennlp/data/token_indexers/elmo_indexer.py
@@ -4,7 +4,7 @@ from overrides import overrides
 import torch
 
 from allennlp.common.util import pad_sequence_to_length
-from allennlp.data.tokenizers.token import Token
+from allennlp.data.tokenizers import Token
 from allennlp.data.token_indexers.token_indexer import TokenIndexer, IndexedTokenList
 from allennlp.data.vocabulary import Vocabulary
