+++ allennlp/data/tokenizers/__init__.py
@@ -3,7 +3,8 @@ This module contains various classes for performing
 tokenization.
 """
 
-from allennlp.data.tokenizers.tokenizer import Token, Tokenizer
+from allennlp.data.tokenizers.token_class import Token
+from allennlp.data.tokenizers.tokenizer import Tokenizer
 from allennlp.data.tokenizers.spacy_tokenizer import SpacyTokenizer
 from allennlp.data.tokenizers.letters_digits_tokenizer import LettersDigitsTokenizer
 from allennlp.data.tokenizers.pretrained_transformer_tokenizer import PretrainedTransformerTokenizer
