+++ allennlp/data/tokenizers/character_tokenizer.py
@@ -2,7 +2,7 @@ from typing import List, Union
 
 from overrides import overrides
 
-from allennlp.data.tokenizers.token import Token
+from allennlp.data.tokenizers.token_class import Token
 from allennlp.data.tokenizers.tokenizer import Tokenizer
