+++ allennlp/data/tokenizers/letters_digits_tokenizer.py
@@ -3,7 +3,7 @@ from typing import List
 
 from overrides import overrides
 
-from allennlp.data.tokenizers.token import Token
+from allennlp.data.tokenizers.token_class import Token
 from allennlp.data.tokenizers.tokenizer import Tokenizer
