+++ tests/data/dataset_readers/dataset_utils/span_utils_test.py
@@ -3,8 +3,7 @@ import pytest
 
 from allennlp.common.testing import AllenNlpTestCase
 from allennlp.data.dataset_readers.dataset_utils import span_utils
-from allennlp.data.tokenizers.spacy_tokenizer import SpacyTokenizer
-from allennlp.data.tokenizers.token import Token
+from allennlp.data.tokenizers import Token, SpacyTokenizer
 
 
 class SpanUtilsTest(AllenNlpTestCase):
